Moved menus and buttons to their own folder
