+++ LegendOfZelda/GameState/Button/RetryButtonBlack.cs
@@ -1,4 +1,5 @@
 ﻿using LegendOfZelda.Interface;
+using LegendOfZelda.Menu;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 namespace LegendOfZelda.GameState.Button
