+++ LegendOfZelda/GameState/Button/SwordInventoryButton.cs
@@ -1,4 +1,6 @@
-﻿using LegendOfZelda.Interface;
+﻿using LegendOfZelda.GameState.Utilities;
+using LegendOfZelda.Interface;
+using LegendOfZelda.Menu;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
