--- LegendOfZelda/GameState/IButtonMenu.cs
@@ -1,10 +1,0 @@
-﻿using LegendOfZelda.GameState.Button;
-using System.Collections.Generic;
-
-namespace LegendOfZelda.GameState
-{
-    internal interface IButtonMenu : IMenu
-    {
-        List<IButton> Buttons { get; }
-    }
-}
