+++ LegendOfZelda/Menu/IButtonMenu.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace LegendOfZelda.Menu
+{
+    internal interface IButtonMenu : IMenu
+    {
+        List<IButton> Buttons { get; }
+    }
+}
